+++ fraudcrawler/base/base.py
@@ -11,8 +11,7 @@ from urllib.parse import urlparse
 import re
 from typing import Any, Dict, List
 
-from bs4 import BeautifulSoup
-from bs4.element import Tag
+
 import httpx
 
 from fraudcrawler.settings import (
@@ -24,7 +23,6 @@ from fraudcrawler.settings import (
     DEFAULT_HTTPX_LIMITS,
     DEFAULT_HTTPX_REDIRECTS,
 )
-from fraudcrawler.settings import TOPPREISE_SEARCH_PATHS, TOPPREISE_COMPARISON_PATHS
 
 logger = logging.getLogger(__name__)
 
@@ -244,108 +242,3 @@ class DomainUtils:
         if hostname and hostname.startswith("www."):
             hostname = hostname[4:]
         return hostname.lower()
-
-
-class ToppreiseUtils:
-    """Utility class for Toppreise specific URL extraction."""
-
-    _endpoint = "https://www.toppreise.ch/"
-    _headers = {
-        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36",
-        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
-        "Accept-Language": "en-US,en;q=0.5",
-        "Accept-Encoding": "gzip, deflate",
-        "Connection": "keep-alive",
-        "Upgrade-Insecure-Requests": "1",
-    }
-
-    @classmethod
-    def _get_search_endpoint(cls, language: Language) -> str:
-        """Get the search endpoint based on the language."""
-        search_path = TOPPREISE_SEARCH_PATHS.get(
-            language.code, TOPPREISE_SEARCH_PATHS["default"]
-        )
-        return f"{cls._endpoint}{search_path}"
-
-    @staticmethod
-    def _extract_links(
-        element: Tag, ext_products: bool = True, comp_products: bool = True
-    ) -> List[str]:
-        """Extracts all relevant product URLs from a BeautifulSoup object of a Toppreise page.
-
-        Note:
-            Depending on the arguments, it extracts:
-                - product comparison URLs (i.e. https://www.toppreise.ch/preisvergleich/...)
-                - external product URLs (i.e. https://www.example.com/ext_...).
-
-        Args:
-            tag: BeautifulSoup Tag object containing the HTML to parse.
-            ext_products: Whether to extract external product URLs.
-            comp_products: Whether to extract product comparison URLs.
-        """
-        # Find all links in the page
-        links = element.find_all("a", href=True)
-
-        # Filter links to only include external product links
-        hrefs = [
-            href
-            for link in links
-            if (
-                hasattr(link, "get")  # Ensure we have a Tag object with href attribute
-                and (href := link.get("href"))  # Ensure href is not None
-                and not href.startswith("javascript:")  # Skip javascript links
-                and isinstance(href, str)  # Ensure href is a string
-                # Make sure the link is either an external product link (href contains 'ext_')
-                # or is a search result link (href contains 'preisvergleich', 'comparison-prix', or 'price-comparison')
-                and (
-                    ("ext_" in href and ext_products)
-                    or (
-                        any(pth in href for pth in TOPPREISE_COMPARISON_PATHS)
-                        and comp_products
-                    )
-                )
-            )
-        ]
-
-        # Make relative URLs absolute
-        urls = []
-        for href in hrefs:
-            if href.startswith("/"):
-                href = f"https://www.toppreise.ch{href}"
-            elif not href.startswith("http"):
-                href = f"https://www.toppreise.ch/{href}"
-            urls.append(href)
-
-        # Return deduplicated urls
-        urls = list(set(urls))
-        return urls
-
-    def _extract_product_urls_from_search_page(self, content: bytes) -> List[str]:
-        """Extracts product urls from a Toppreise search page (i.e. https://www.toppreise.ch/produktsuche)."""
-
-        # Parse the HTML
-        soup = BeautifulSoup(content, "html.parser")
-        main = soup.find("div", id="Page_Browsing")
-        if not isinstance(main, Tag):
-            logger.warning("No main content found in Toppreise search page.")
-            return []
-
-        # Extract links (external product links and comparison links)
-        urls = self._extract_links(element=main)
-
-        logger.debug(f"Found {len(urls)} product URLs from Toppreise search results.")
-        return urls
-
-    def _extract_product_urls_from_comparison_page(self, content: bytes) -> List[str]:
-        """Extracts product urls from a Toppreise product comparison page (i.e. https://www.toppreise.ch/preisvergleich/...)."""
-
-        # Parse the HTML
-        soup = BeautifulSoup(content, "html.parser")
-
-        # Extract links (external product links only)
-        urls = self._extract_links(element=soup, comp_products=False)
-
-        logger.debug(
-            f"Found {len(urls)} external product URLs from Toppreise comparison page."
-        )
-        return urls
