+++ tests/unittests/test_base.py
@@ -9,9 +9,7 @@ from fraudcrawler.base.base import (
     Deepness,
     Prompt,
     DomainUtils,
-    ToppreiseUtils,
 )
-from fraudcrawler.settings import ROOT_DIR
 
 
 def test_setup():
@@ -125,45 +123,3 @@ def test_prompt():
 def test_domain_utils_get_domain(url, expected):
     du = DomainUtils()
     assert du._get_domain(url) == expected
-
-
-def test_toppreise_utils_get_search_endpoint():
-    language = Language(name="German", code="de")
-    endpoint = ToppreiseUtils._get_search_endpoint(language=language)
-    assert endpoint == "https://www.toppreise.ch/produktsuche"
-
-    language = Language(name="French", code="fr")
-    endpoint = ToppreiseUtils._get_search_endpoint(language=language)
-    assert endpoint == "https://www.toppreise.ch/chercher"
-
-    language = Language(name="English", code="en")
-    endpoint = ToppreiseUtils._get_search_endpoint(language=language)
-    assert endpoint == "https://www.toppreise.ch/browse"
-
-
-def test_toppreise_utils_extract_product_urls_from_search_page():
-    tu = ToppreiseUtils()
-    with open(ROOT_DIR / "tests" / "data" / "toppreise_search.html", "rb") as f:
-        content = f.read()
-    urls = tu._extract_product_urls_from_search_page(content=content)
-    assert len(urls) == 23
-    assert (
-        "https://www.toppreise.ch/preisvergleich/Kuehl-Gefrierkombinationen/LIEBHERR-CT-2531-p615781?selsort=rd"
-        in urls
-    )
-    assert (
-        "https://www.toppreise.ch/ext_de?pid=0&did=2511&oid=506961161&gdt=MjAyNS0wOS0wNCAyMjo0Mjo0OQ==&slsrt=rd&prcst=shipping&lpos=10"
-        in urls
-    )
-
-
-def test_toppreise_utils_extract_product_urls_from_comparison_page():
-    tu = ToppreiseUtils()
-    with open(ROOT_DIR / "tests" / "data" / "toppreise_comparison.html", "rb") as f:
-        content = f.read()
-    urls = tu._extract_product_urls_from_comparison_page(content=content)
-    assert len(urls) == 20
-    assert (
-        "https://www.toppreise.ch/ext_de?pid=615781&did=2532&oid=493842592&gdt=MjAyNS0wOS0wNCAyMjo0NDoyMw==&slsrt=pa&prcst=shipping&lpos=5"
-        in urls
-    )
